pyproject.toml
[project]
name = "ee641-hw1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
